removed .history files

+++ custom/gov/types/keys_test.go
@@ -17,19 +17,19 @@ func TestProposalKeys(t *testing.T) {
 	// key proposal
 	key := govtypes.ProposalKey(1)
 	proposalID := govtypes.SplitProposalKey(key)
-	require.Equal(t, int(proposalID), 1)
+	require.Equal(t, proposalID, uint64(1))
 
 	// key active proposal queue
 	now := time.Now()
 	key = govtypes.ActiveProposalQueueKey(3, now)
 	proposalID, expTime := govtypes.SplitActiveProposalQueueKey(key)
-	require.Equal(t, int(proposalID), 3)
+	require.Equal(t, proposalID, uint64(3))
 	require.True(t, now.Equal(expTime))
 
 	// key inactive proposal queue
 	key = govtypes.InactiveProposalQueueKey(3, now)
 	proposalID, expTime = govtypes.SplitInactiveProposalQueueKey(key)
-	require.Equal(t, int(proposalID), 3)
+	require.Equal(t, proposalID, uint64(3))
 	require.True(t, now.Equal(expTime))
 
 	// invalid key
@@ -40,27 +40,27 @@ func TestProposalKeys(t *testing.T) {
 func TestDepositKeys(t *testing.T) {
 	key := govtypes.DepositsKey(2)
 	proposalID := govtypes.SplitProposalKey(key)
-	require.Equal(t, int(proposalID), 2)
+	require.Equal(t, proposalID, uint64(2))
 
 	key = govtypes.DepositKey(2, addr)
 	proposalID, depositorAddr := govtypes.SplitKeyDeposit(key)
-	require.Equal(t, int(proposalID), 2)
+	require.Equal(t, proposalID, uint64(2))
 	require.Equal(t, addr, depositorAddr)
 }
 
 func TestVoteKeys(t *testing.T) {
 	key := govtypes.VotesKey(2)
 	proposalID := govtypes.SplitProposalKey(key)
-	require.Equal(t, int(proposalID), 2)
+	require.Equal(t, proposalID, uint64(2))
 
 	key = govtypes.VoteKey(2, addr)
 	proposalID, voterAddr := govtypes.SplitKeyDeposit(key)
-	require.Equal(t, int(proposalID), 2)
+	require.Equal(t, proposalID, uint64(2))
 	require.Equal(t, addr, voterAddr)
 }
 
 func TestTotalDepositKeys(t *testing.T) {
 	key := TotalDepositKey(2)
 	proposalID := govtypes.SplitProposalKey(key)
-	require.Equal(t, int(proposalID), 2)
+	require.Equal(t, proposalID, uint64(2))
 }
